feat(code): add board view mode to inbox
Adds a "View as" toggle to the inbox toolbar with a list (default) and board option. The board groups reports into kanban columns by status (Ready, Needs input, Researching, Failed, Queued, Gathering), respecting active status filters. Selecting a card opens the standard detail pane in a right-side panel. The view mode is persisted in the filter store.

Generated-By: PostHog Code
Task-Id: bb7bd45f-d16b-4cca-a8bd-00856f6c5004
